Added tick counters to track performance of network library. Removed need to checksum ICMP reply packets. Saved 3ms on 1200byte pings and 100 bytes of code

+++ CubeV2/CPUTime.h
@@ -6,13 +6,13 @@
 */
 #include <avr/io.h>
 #include <stdint.h>
-
+#include <util/atomic.h>
 #ifndef CPUTIME_H_
 #define CPUTIME_H_
 
-extern volatile uint32_t netTick;
+extern volatile uint64_t netTick;
 
-uint32_t GetTicks();
+uint64_t GetTicks();
 
 typedef struct {
 	uint32_t Ticks;
